refactor: moved auth service to dedicated folder
