+++ frontend/src/app/core/auth/components/login/login.component.ts
@@ -4,7 +4,7 @@ import { ReactiveFormsModule, FormGroup, FormControl, Validators } from '@angula
 import { Router } from '@angular/router';
 import { HttpErrorResponse } from '@angular/common/http';
 
-import { AuthService } from '../../services/auth.service';
+import { AuthService } from '../../services/auth/auth.service';
 import { AlertComponent } from '../../../../shared/components/alert/alert.component';
 
 @Component({
