+++ frontend/src/app/core/guards/auth/auth.guard.ts
@@ -1,7 +1,7 @@
 import { Injectable, inject } from "@angular/core";
 import { CanActivateFn, Router } from "@angular/router";
 
-import { AuthService } from '../../auth/services/auth.service'
+import { AuthService } from '../../auth/services/auth/auth.service'
 
 @Injectable({ providedIn: 'root' })
